+++ ft_strncmp.c
@@ -6,7 +6,7 @@
 /*   By: amzahir <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/26 16:24:35 by amzahir           #+#    #+#             */
-/*   Updated: 2024/10/27 21:14:48 by amzahir          ###   ########.fr       */
+/*   Updated: 2024/11/08 17:09:29 by amzahir          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,19 +14,18 @@
 
 int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	unsigned long int	i;
+	size_t	i;
 	
 	i = 0;
-	if(!n)
-		return(0);
+	if (!n)
+		return (0);
 	s1 = (unsigned char*) s1;
 	s2 = (unsigned char*) s2;
-	while(s1[i] && s2[i] && i < n - 1)
+	while (s1[i] && s2[i] && i < n - 1)
 	{
 		if (s1[i] != s2[i])
-			//return(s1[i] - s2[i]);
 			break;
 		i++;
 	}
-	return(s1[i] - s2[i]);
+	return (s1[i] - s2[i]);
 }
